Devs

+++ street-goods-burger-backend/app/models/concerns/store_admin_features.rb
@@ -1,28 +1,47 @@
-# frozen_string_literal: true
+# git push --set-upstream origin fix/feature/store_admin/view_store_transactions           frozen_string_literal: true
 
 # StoreAdminFeatures module
 module StoreAdminFeatures
   def view_store_transactions(store_id:)
     store_transactions = StoreTransaction.where(store_id:).map do |transaction|
       store_transaction_id = transaction[:id]
+      transaction.cart
 
-      # get store transaction corressponding cart
-      cart_info = JSON.parse(Cart.where(store_transaction_id:).to_json)[0]
-      cart_info_id = cart_info['id']
+      if transaction.cart.nil?
 
-      {
-        id: store_transaction_id,
-        status: transaction[:status],
-        store_customer_email: transaction.store_customer[:email],
-        created_at: transaction[:created_at],
-        updated_at: transaction[:updated_at],
-        cart_info: {
-          id: cart_info_id,
-          total_price: cart_info['total_price'],
-          quantity: cart_info['quantity']
+        # get store transaction corressponding cart
+
+        {
+          id: store_transaction_id,
+          status: transaction[:status],
+          store_customer_email: transaction.store_customer[:email],
+          created_at: transaction[:created_at],
+          updated_at: transaction[:updated_at],
+          cart_info: {
+            message: 'no cart yet'
+          }
         }
-      }
+
+      else
+        # get store transaction corressponding cart
+        cart_info = transaction.cart
+        cart_info_id = cart_info[:id]
+
+        {
+          id: store_transaction_id,
+          status: transaction[:status],
+          store_customer_email: transaction.store_customer[:email],
+          created_at: transaction[:created_at],
+          updated_at: transaction[:updated_at],
+          cart_info: {
+            id: cart_info_id,
+            total_price: cart_info['total_price'],
+            quantity: cart_info['quantity']
+          }
+        }
+      end
     end
+
     {
       status: 200,
       message: 'successfully fetch store transactions',
@@ -103,7 +122,12 @@ module StoreAdminFeatures
   end
 
   def view_store_customer_food_orders(store_customer_id:)
-    food_orders = StoreCustomer.find(store_customer_id).store_transactions.all.map do |store_transaction|
+    # store transaction with cart
+    with_cart = StoreCustomer.find(store_customer_id).store_transactions.all.filter do |store_transaction|
+      !store_transaction.cart.nil?
+    end
+
+    food_orders = with_cart.map do |store_transaction|
       store_transaction.cart.food_orders
     end
 
@@ -116,19 +140,19 @@ module StoreAdminFeatures
         food_orders:
       }
     }
-  rescue ActiveRecord::RecordNotFound => e
+    rescue ActiveRecord::RecordNotFound => e
     {
       status: 400,
       message: 'invalid data provided',
       error: e.message
     }
-  rescue ActiveRecord::RecordInvalid => e
+    rescue ActiveRecord::RecordInvalid => e
     {
       status: 422,
       message: 'invalid data provided',
       error: e.message
     }
-  rescue NoMethodError => e
+    rescue NoMethodError => e
     {
       status: 422,
       message: "can't do calculations based on data provided",
@@ -321,9 +345,7 @@ module StoreAdminFeatures
 
   def remove_food(food_id:)
     # if food has reviews delete reviews first
-    unless (store.foods.find(food_id).reviews.length.zero?) 
-      store.foods.find(food_id).reviews.where(food_id:).delete_all
-    end
+    store.foods.find(food_id).reviews.where(food_id:).delete_all unless store.foods.find(food_id).reviews.length.zero?
 
     food = store.foods.find(food_id).delete
     {
